build: update dependencies and use kotlin test due to build fail
