+++ gradle.properties
@@ -3,3 +3,5 @@ kotlin.native.cacheKind.linuxX64=none
 org.gradle.caching=true
 org.gradle.jvmargs=-Dfile.encoding=UTF-8
 org.gradle.parallel=true
+org.jetbrains.dokka.experimental.gradle.pluginMode=V2Enabled
+org.jetbrains.dokka.experimental.gradle.pluginMode.noWarn=true
