+++ gradle/libs.versions.toml
@@ -1,24 +1,24 @@
 [versions]
-detekt = "1.23.5"
-dokka = "1.9.20"
-kotest = "5.8.0"
-kotlin = "1.9.23"
-kotlinxBinaryCompatibilityValidator = "0.14.0"
-kotlinxSerialization = "1.6.3"
-kover = "0.7.6"
+burst = "2.4.0"
+detekt = "1.23.8"
+dokka = "2.0.0"
+kotest = "5.9.1"
+kotlin = "2.1.10"
+kotlinxBinaryCompatibilityValidator = "0.17.0"
+kotlinxSerialization = "1.8.0"
+kover = "0.9.1"
 
 [libraries]
 kotest-assertionsCore = { module = "io.kotest:kotest-assertions-core", version.ref = "kotest" }
-kotest-frameworkDatatest = { module = "io.kotest:kotest-framework-datatest", version.ref = "kotest" }
-kotest-frameworkEngine = { module = "io.kotest:kotest-framework-engine", version.ref = "kotest" }
-kotest-runnerJunit5 = { module = "io.kotest:kotest-runner-junit5", version.ref = "kotest" }
+kotlin-test = { module = "org.jetbrains.kotlin:kotlin-test", version.ref = "kotlin" }
+kotlin-test-junit5 = { module = "org.jetbrains.kotlin:kotlin-test-junit5", version.ref = "kotlin" }
 kotlinx-serialization-core = { module = "org.jetbrains.kotlinx:kotlinx-serialization-core", version.ref = "kotlinxSerialization" }
 kotlinx-serialization-json = { module = "org.jetbrains.kotlinx:kotlinx-serialization-json", version.ref = "kotlinxSerialization" }
 
 [plugins]
+burst = { id = "app.cash.burst", version.ref = "burst" }
 detekt = { id = "io.gitlab.arturbosch.detekt", version.ref = "detekt" }
 dokka = { id = "org.jetbrains.dokka", version.ref = "dokka" }
-kotest-multiplatform = { id = "io.kotest.multiplatform", version.ref = "kotest" }
 kotlin-multiplatform = { id = "org.jetbrains.kotlin.multiplatform", version.ref = "kotlin" }
 kotlinx-binaryCompatibilityValidator = { id = "org.jetbrains.kotlinx.binary-compatibility-validator", version.ref = "kotlinxBinaryCompatibilityValidator" }
 kotlinx-serialization = { id = "org.jetbrains.kotlin.plugin.serialization", version.ref = "kotlin" }
