--- src/commonTest/kotlin/net/swiftzer/semver/SemVerParseFailTest.kt
@@ -1,63 +1,0 @@
-package net.swiftzer.semver
-
-import io.kotest.assertions.throwables.shouldThrow
-import io.kotest.core.spec.style.FunSpec
-import io.kotest.datatest.withData
-import io.kotest.matchers.nulls.shouldBeNull
-
-@Suppress("MaxLineLength")
-class SemVerParseFailTest : FunSpec(
-    {
-        withData(
-            nameFn = { "fail case [$it]" },
-            ts = sequenceOf(
-                "v1.2.3",
-                " 1.2.3",
-                "1.2.3 ",
-                "1",
-                "a",
-                "1.2",
-                "1.2.3-0123",
-                "1.2.3-0123.0123",
-                "1.1.2+.123",
-                "+invalid",
-                "-invalid",
-                "-invalid+invalid",
-                "-invalid.01",
-                "alpha",
-                "alpha.beta",
-                "alpha.beta.1",
-                "alpha.1",
-                "alpha+beta",
-                "alpha_beta",
-                "alpha.",
-                "alpha..",
-                "beta",
-                "1.0.0-alpha_beta",
-                "-alpha.",
-                "1.0.0-alpha..",
-                "1.0.0-alpha..1",
-                "1.0.0-alpha...1",
-                "1.0.0-alpha....1",
-                "1.0.0-alpha.....1",
-                "1.0.0-alpha......1",
-                "1.0.0-alpha.......1",
-                "01.1.1",
-                "1.01.1",
-                "1.1.01",
-                "1.2.3.DEV",
-                "1.2-SNAPSHOT",
-                "1.2.31.2.3----RC-SNAPSHOT.12.09.1--..12+788",
-                "1.2-RC-SNAPSHOT",
-                "-1.0.3-gamma+b7718",
-                "+justmeta",
-                "9.8.7+meta+meta",
-                "9.8.7-whatever+meta+meta",
-                "99999999999999999999999.999999999999999999.99999999999999999----RC-SNAPSHOT.12.09.1--------------------------------..12",
-            ),
-        ) {
-            shouldThrow<IllegalArgumentException> { SemVer.parse(it) }
-            SemVer.parseOrNull(it).shouldBeNull()
-        }
-    },
-)
